+++ mobile/src/main/java/net/mm2d/dmsexplorer/settings/theme/ServerColorExtractorNormal.java
@@ -5,7 +5,7 @@
  * http://opensource.org/licenses/MIT
  */
 
-package net.mm2d.dmsexplorer.util;
+package net.mm2d.dmsexplorer.settings.theme;
 
 import android.graphics.Bitmap;
 import android.support.annotation.NonNull;
@@ -15,11 +15,12 @@ import android.support.v7.graphics.Palette.Swatch;
 
 import net.mm2d.android.upnp.cds.MediaServer;
 import net.mm2d.dmsexplorer.Const;
+import net.mm2d.dmsexplorer.util.PaletteUtils;
 
 /**
  * @author <a href="mailto:ryo@mm2d.net">大前良介 (OHMAE Ryosuke)</a>
  */
-public class ServerColorExtractorDark implements ServerColorExtractor {
+public class ServerColorExtractorNormal implements ServerColorExtractor {
     private static final ThemeColorGenerator GENERATOR = new ThemeColorGeneratorDark();
 
     @Override
@@ -62,11 +63,11 @@ public class ServerColorExtractorDark implements ServerColorExtractor {
         if (palette != null) {
             final Swatch lightSwatch = PaletteUtils.selectLightSwatch(palette);
             if (lightSwatch != null) {
-                expandedColor = ColorUtils.getDarkerColor(lightSwatch.getRgb(), 0.3f);
+                expandedColor = lightSwatch.getRgb();
             }
             final Swatch darkSwatch = PaletteUtils.selectDarkSwatch(palette);
             if (darkSwatch != null) {
-                collapsedColor = ColorUtils.getDarkerColor(darkSwatch.getRgb(), 0.3f);
+                collapsedColor = darkSwatch.getRgb();
             }
         }
         server.putIntTag(Const.KEY_TOOLBAR_EXPANDED_COLOR, expandedColor);
