+++ mobile/src/main/java/net/mm2d/dmsexplorer/settings/theme/ThemeColorGeneratorNormal.java
@@ -5,7 +5,7 @@
  * http://opensource.org/licenses/MIT
  */
 
-package net.mm2d.dmsexplorer.util;
+package net.mm2d.dmsexplorer.settings.theme;
 
 import android.graphics.Color;
 import android.support.annotation.NonNull;
@@ -14,14 +14,14 @@ import android.text.TextUtils;
 /**
  * @author <a href="mailto:ryo@mm2d.net">大前良介 (OHMAE Ryosuke)</a>
  */
-public class ThemeColorGeneratorDark implements ThemeColorGenerator {
+public class ThemeColorGeneratorNormal implements ThemeColorGenerator {
     @Override
     public int getIconColor(@NonNull final String title) {
         final char c = TextUtils.isEmpty(title) ? ' ' : title.charAt(0);
         return Color.HSVToColor(new float[]{
                 (59 * c) % 360,
-                141f / 255f,
-                83f / 255f,
+                111f / 255f,
+                237f / 255f,
         });
     }
 
@@ -31,7 +31,7 @@ public class ThemeColorGeneratorDark implements ThemeColorGenerator {
         return Color.HSVToColor(new float[]{
                 (59 * c) % 360,
                 185f / 255f,
-                32f / 255f,
+                187f / 255f,
         });
     }
 
@@ -41,7 +41,7 @@ public class ThemeColorGeneratorDark implements ThemeColorGenerator {
         return Color.HSVToColor(new float[]{
                 (59 * c) % 360,
                 124f / 255f,
-                48f / 255f,
+                210f / 255f,
         });
     }
 
@@ -50,8 +50,8 @@ public class ThemeColorGeneratorDark implements ThemeColorGenerator {
         final char c = TextUtils.isEmpty(title) ? ' ' : title.charAt(0);
         return Color.HSVToColor(new float[]{
                 (59 * c) % 360,
-                185f / 255f,
-                48f / 255f,
+                34f / 255f,
+                248f / 255f,
         });
     }
 
@@ -61,7 +61,7 @@ public class ThemeColorGeneratorDark implements ThemeColorGenerator {
         return Color.HSVToColor(new float[]{
                 (59 * c) % 360,
                 192f / 255f,
-                32f / 255f,
+                96f / 255f,
         });
     }
 }
