+++ mobile/src/main/java/net/mm2d/dmsexplorer/viewmodel/ServerDetailFragmentModel.java
@@ -27,7 +27,7 @@ import net.mm2d.dmsexplorer.R;
 import net.mm2d.dmsexplorer.Repository;
 import net.mm2d.dmsexplorer.log.EventLogger;
 import net.mm2d.dmsexplorer.settings.Settings;
-import net.mm2d.dmsexplorer.util.ThemeColorGenerator;
+import net.mm2d.dmsexplorer.settings.theme.ThemeColorGenerator;
 import net.mm2d.dmsexplorer.view.ContentListActivity;
 import net.mm2d.dmsexplorer.view.adapter.PropertyAdapter;
 import net.mm2d.upnp.Icon;
