+++ mobile/src/main/java/net/mm2d/dmsexplorer/viewmodel/ServerItemModel.java
@@ -22,8 +22,8 @@ import net.mm2d.android.util.DrawableUtils;
 import net.mm2d.dmsexplorer.R;
 import net.mm2d.dmsexplorer.settings.ColorThemeParams;
 import net.mm2d.dmsexplorer.settings.Settings;
-import net.mm2d.dmsexplorer.util.ServerColorExtractor;
-import net.mm2d.dmsexplorer.util.ThemeColorGenerator;
+import net.mm2d.dmsexplorer.settings.theme.ServerColorExtractor;
+import net.mm2d.dmsexplorer.settings.theme.ThemeColorGenerator;
 import net.mm2d.upnp.Icon;
 
 /**
